+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
+### Added
+- `DiscoApp::Configuration` model, to be set up with `disco_app.rb` initializer
+
 ### Changed
 - All `id` fields default to `bigint` thanks to `rails-bigint-pk` gem