+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
+No unreleased changes.
+
+## 0.8.4 - 2016-02-22
 ### Added
 - `DiscoApp::Configuration` model, to be set up with `disco_app.rb` initializer
 - Automatically set `@shop` instance variable on Proxy Controller concern